requirements.txt

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning and Deep Learning
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Image Processing and Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0

# Medical Imaging
nibabel>=3.2.0

# Visualization
matplotlib>=3.4.0
plotly>=5.0.0
seaborn>=0.11.0

# Web Framework
Flask>=2.0.0
Werkzeug>=2.0.0

# Dashboard and Interactive Components
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Progress Tracking
tqdm>=4.62.0

# Development and Logging
tensorboard>=2.7.0

# Additional utilities (commonly needed)
requests>=2.25.0
Jinja2>=3.0.0
